+++ controllers/resource/volumeGenerator.go
@@ -176,6 +176,15 @@ func createStaticVolumes(doguResource *k8sv2.Dogu) []corev1.Volume {
 		},
 	}
 
+	hostTimezoneVolume := corev1.Volume{
+		Name: localTimeMountName,
+		VolumeSource: corev1.VolumeSource{
+			HostPath: &corev1.HostPathVolumeSource{
+				Path: localTimeMountPath,
+			},
+		},
+	}
+
 	// add EmptyDir-VolumeSource for all dogus to at least give them the ability to write state
 	ephemeralVolume := corev1.Volume{
 		Name: doguResource.GetEphemeralDataVolumeName(),
@@ -219,6 +228,7 @@ func createStaticVolumes(doguResource *k8sv2.Dogu) []corev1.Volume {
 		globalConfigVolume,
 		normalConfigVolume,
 		sensitiveConfigVolume,
+		hostTimezoneVolume,
 	}
 }
 
@@ -330,6 +340,10 @@ func createStaticVolumeMounts(doguResource *k8sv2.Dogu) []corev1.VolumeMount {
 			Name:      globalConfig,
 			ReadOnly:  true,
 			MountPath: "/etc/ces/config/global",
+		}, {
+			Name:      localTimeMountName,
+			ReadOnly:  true,
+			MountPath: localTimeMountPath,
 		},
 	}